+++ src/lorawan/encoding.hpp
@@ -3,8 +3,9 @@
 #define ENCODING_HPP
 
 #include <stdint.h>
+#include <datatypes.hpp>
 
-uint32_t encode_ambLight();
+uint24 encode_ambLight();
 uint32_t encode_envSensor();
 
 #endif // ENCODING_HPP
